+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Added several cluster-wide WaitConditions for Deployments, DaemonSets, StatefulSets, Jobs and Pods being in an expected state.
 
+### Fixed
+
+- Fallback to the default context in the Kubeconfig if empty context name provided
+
 ## [1.24.1] - 2024-09-19
 
 ### Fixed